requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
click==8.1.3
    # via uvicorn
django==4.1.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.13.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
pytz==2022.4
    # via djangorestframework
sqlparse==0.4.3
    # via django
uvicorn==0.18.3
    # via -r requirements.in
